Update from gnulib.

+++ lib/hash.h
@@ -25,8 +25,9 @@
 # define HASH_H_
 
 # include <stdio.h>
+# include <stdbool.h>
 
-typedef unsigned (*Hash_hasher) (const void *, unsigned);
+typedef size_t (*Hash_hasher) (const void *, size_t);
 typedef bool (*Hash_comparator) (const void *, const void *);
 typedef void (*Hash_data_freer) (void *);
 typedef bool (*Hash_processor) (void *, void *);
@@ -56,10 +57,10 @@ struct hash_table;
 typedef struct hash_table Hash_table;
 
 /* Information and lookup.  */
-unsigned hash_get_n_buckets (const Hash_table *);
-unsigned hash_get_n_buckets_used (const Hash_table *);
-unsigned hash_get_n_entries (const Hash_table *);
-unsigned hash_get_max_bucket_length (const Hash_table *);
+size_t hash_get_n_buckets (const Hash_table *);
+size_t hash_get_n_buckets_used (const Hash_table *);
+size_t hash_get_n_entries (const Hash_table *);
+size_t hash_get_max_bucket_length (const Hash_table *);
 bool hash_table_ok (const Hash_table *);
 void hash_print_statistics (const Hash_table *, FILE *);
 void *hash_lookup (const Hash_table *, const void *);
@@ -67,20 +68,20 @@ void *hash_lookup (const Hash_table *, const void *);
 /* Walking.  */
 void *hash_get_first (const Hash_table *);
 void *hash_get_next (const Hash_table *, const void *);
-unsigned hash_get_entries (const Hash_table *, void **, unsigned);
-unsigned hash_do_for_each (const Hash_table *, Hash_processor, void *);
+size_t hash_get_entries (const Hash_table *, void **, size_t);
+size_t hash_do_for_each (const Hash_table *, Hash_processor, void *);
 
 /* Allocation and clean-up.  */
-unsigned hash_string (const char *, unsigned);
+size_t hash_string (const char *, size_t);
 void hash_reset_tuning (Hash_tuning *);
-Hash_table *hash_initialize (unsigned, const Hash_tuning *,
+Hash_table *hash_initialize (size_t, const Hash_tuning *,
 			     Hash_hasher, Hash_comparator,
 			     Hash_data_freer);
 void hash_clear (Hash_table *);
 void hash_free (Hash_table *);
 
 /* Insertion and deletion.  */
-bool hash_rehash (Hash_table *, unsigned);
+bool hash_rehash (Hash_table *, size_t);
 void *hash_insert (Hash_table *, const void *);
 void *hash_delete (Hash_table *, const void *);
